+++ app/pages/publications/index.js
@@ -8,6 +8,7 @@ import createPublicationOnCurriculum from "app/publication-on-curricula/mutation
 import InformationCard from "app/core/components/InformationCard"
 import { Grid, Button, Chip, Select, MenuItem, InputLabel, FormControl } from "@mui/material"
 import CustomSpinner from "app/core/components/CustomSpinner"
+import Swal from "sweetalert2"
 const ITEMS_PER_PAGE = 100
 export const PublicationsList = (props) => {
   const router = useRouter()
@@ -68,21 +69,36 @@ export const PublicationsList = (props) => {
   }
 
   const handleOnDelete = async (id) => {
-    if ((props.curriculumId !== undefined && props.curriculumId !== "") || props.onCurriculum) {
-      await deletePublicationOnCurriculumMutation({
-        curriculumId: props.curriculumId,
-        publicationId: id,
-      })
-      const newPublications = publicationsList.filter((publication) => publication.id !== id)
-      setPublicationsList(newPublications)
-      const newOptions = [...options, allPublications.find((publication) => publication.id === id)]
-      setOptions(newOptions)
-    } else {
-      await deletePublicationMutation({
-        id,
-      })
-      router.push(Routes.PublicationsPage())
-    }
+    Swal.fire({
+      title: "La publicación se eliminará, esta seguro?",
+      showDenyButton: true,
+      confirmButtonText: "Eliminar",
+      denyButtonText: `No eliminar`,
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        Swal.fire("La publicación se elimino", "", "info")
+        if ((props.curriculumId !== undefined && props.curriculumId !== "") || props.onCurriculum) {
+          await deletePublicationOnCurriculumMutation({
+            curriculumId: props.curriculumId,
+            publicationId: id,
+          })
+          const newPublications = publicationsList.filter((publication) => publication.id !== id)
+          setPublicationsList(newPublications)
+          const newOptions = [
+            ...options,
+            allPublications.find((publication) => publication.id === id),
+          ]
+          setOptions(newOptions)
+        } else {
+          await deletePublicationMutation({
+            id,
+          })
+          router.push(Routes.PublicationsPage())
+        }
+      } else if (result.isDenied) {
+        Swal.fire("La publicación no se elimino", "", "info")
+      }
+    })
   }
 
   return (
